Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/selection/deleteFromDocument-expected.txt

To debug test failures, add a query parameter with the test id (like "?5"). Only that test will be run. Then you can look at the resulting iframes in the DOM.


PASS Selection.prototype.deleteFromDocument.length must equal 0
PASS Range 0: empty
PASS Range 1: [paras[0].firstChild, 0, paras[0].firstChild, 0]
FAIL Range 2: [paras[0].firstChild, 0, paras[0].firstChild, 1] assert_true: First differing node: expected Text node "̈b̈c̈d̈ëf̈g̈ḧ
", got Text node "b̈c̈d̈ëf̈g̈ḧ
" expected true got false
PASS Range 3: [paras[0].firstChild, 2, paras[0].firstChild, 8]
FAIL Range 4: [paras[0].firstChild, 2, paras[0].firstChild, 9] assert_true: First differing node: expected Text node "Ä̈f̈g̈ḧ
", got Text node "Äf̈g̈ḧ
" expected true got false
PASS Range 5: [paras[1].firstChild, 0, paras[1].firstChild, 0]
FAIL Range 6: [paras[1].firstChild, 0, paras[1].firstChild, 1] assert_true: First differing node: expected Text node "jklmnop
", got Text node "Ijklmnop
" expected true got false
FAIL Range 7: [paras[1].firstChild, 2, paras[1].firstChild, 8] assert_true: First differing node: expected Text node "Ij
", got Text node "Ijklmnop
" expected true got false
FAIL Range 8: [paras[1].firstChild, 2, paras[1].firstChild, 9] assert_true: First differing node: expected Text node "Ij", got Text node "Ijklmnop
" expected true got false
PASS Range 9: [detachedPara1.firstChild, 0, detachedPara1.firstChild, 0]
PASS Range 10: [detachedPara1.firstChild, 0, detachedPara1.firstChild, 1]
PASS Range 11: [detachedPara1.firstChild, 2, detachedPara1.firstChild, 8]
PASS Range 12: [foreignPara1.firstChild, 0, foreignPara1.firstChild, 0]
PASS Range 13: [foreignPara1.firstChild, 0, foreignPara1.firstChild, 1]
PASS Range 14: [foreignPara1.firstChild, 2, foreignPara1.firstChild, 8]
FAIL Range 15: [document.documentElement, 0, document.documentElement, 1] assert_true: First differing node: expected Element node <html><body><div id="test"><p id="a">Äb̈c̈d̈ëf̈g̈ḧ
</p..., got Element node <html><head><title>Selection test iframe</title>
<link re... expected true got false
FAIL Range 16: [document.documentElement, 0, document.documentElement, 2] assert_true: First differing node: expected Element node <html></html>, got Element node <html><head><title>Selection test iframe</title>
<link re... expected true got false
FAIL Range 17: [document.documentElement, 1, document.documentElement, 2] assert_true: First differing node: expected Element node <html><head><title>Selection test iframe</title>
<link re..., got Element node <html><head><title>Selection test iframe</title>
<link re... expected true got false
PASS Range 18: [document.head, 1, document.head, 1]
FAIL Range 19: [document.body, 0, document.body, 1] assert_true: First differing node: expected Element node <body>


</body>, got Element node <body><div id="test"><p id="a"></p><p id="c"></p><p id="d... expected true got false
PASS Range 20: [foreignDoc.documentElement, 0, foreignDoc.documentElement, 1]
PASS Range 21: [foreignDoc.head, 1, foreignDoc.head, 1]
PASS Range 22: [foreignDoc.body, 0, foreignDoc.body, 0]
PASS Range 23: [paras[0], 0, paras[0], 0]
FAIL Range 24: [paras[0], 0, paras[0], 1] assert_true: First differing node: expected Element node <p id="a"></p>, got Element node <p id="a">
</p> expected true got false
PASS Range 25: [detachedPara1, 0, detachedPara1, 0]
PASS Range 26: [detachedPara1, 0, detachedPara1, 1]
FAIL Range 27: [paras[0].firstChild, 0, paras[1].firstChild, 0] assert_true: First differing node: expected Element node <div id="test"><p id="a"></p><p id="b" style="display:non..., got Element node <div id="test"><p id="a"></p><p id="c">Qrstuvwx</p><p id=... expected true got false
FAIL Range 28: [paras[0].firstChild, 0, paras[1].firstChild, 8] assert_true: First differing node: expected Element node <div id="test"><p id="a"></p><p id="b" style="display:non..., got Element node <div id="test"><p id="a"></p><p id="c">Qrstuvwx</p><p id=... expected true got false
FAIL Range 29: [paras[0].firstChild, 3, paras[3], 1] assert_true: First differing node: expected Element node <div id="test"><p id="a">Äb</p><p id="d" style="display:..., got Element node <div id="test"><p id="a">Äb̈</p><p id="c"></p><p id="d" ... expected true got false
FAIL Range 30: [paras[0], 0, paras[0].firstChild, 7] assert_true: First differing node: expected Text node "̈ëf̈g̈ḧ
", got Text node "ëf̈g̈ḧ
" expected true got false
FAIL Range 31: [testDiv, 2, paras[4], 1] assert_true: First differing node: expected Element node <div id="test"><p id="a">Äb̈c̈d̈ëf̈g̈ḧ
</p><p id="b" s..., got Element node <div id="test"><p id="a">Äb̈c̈d̈ëf̈g̈ḧ
</p><p id="b" s... expected true got false
FAIL Range 32: [testDiv, 1, paras[2].firstChild, 5] assert_true: First differing node: expected Element node <div id="test"><p id="a">Äb̈c̈d̈ëf̈g̈ḧ
</p><p id="c">v..., got Element node <div id="test"><p id="a">Äb̈c̈d̈ëf̈g̈ḧ
</p><p id="b" s... expected true got false
PASS Range 33: [document.documentElement, 1, document.body, 0]
PASS Range 34: [foreignDoc.documentElement, 1, foreignDoc.body, 0]
FAIL Range 35: [document, 0, document, 1] assert_true: First differing node: expected Document node with 1 child, got Document node with 2 children expected true got false
FAIL Range 36: [document, 0, document, 2] assert_true: First differing node: expected Document node with 0 children, got Document node with 2 children expected true got false
FAIL Range 37: [document, 1, document, 2] assert_true: First differing node: expected Document node with 1 child, got Document node with 2 children expected true got false
FAIL Range 38: [testDiv, 0, comment, 5] assert_true: First differing node: expected Element node <div id="test"><!--bet soup?--></div>, got Element node <div id="test"><p id="a"></p><p id="c"></p><p id="d" styl... expected true got false
FAIL Range 39: [paras[2].firstChild, 4, comment, 2] assert_true: First differing node: expected Element node <div id="test"><p id="a">Äb̈c̈d̈ëf̈g̈ḧ
</p><p id="b" s..., got Element node <div id="test"><p id="a">Äb̈c̈d̈ëf̈g̈ḧ
</p><p id="b" s... expected true got false
FAIL Range 40: [paras[3], 1, comment, 8] assert_true: First differing node: expected Element node <div id="test"><p id="a">Äb̈c̈d̈ëf̈g̈ḧ
</p><p id="b" s..., got Element node <div id="test"><p id="a">Äb̈c̈d̈ëf̈g̈ḧ
</p><p id="b" s... expected true got false
PASS Range 41: [foreignDoc, 0, foreignDoc, 0]
PASS Range 42: [foreignDoc, 1, foreignComment, 2]
PASS Range 43: [foreignDoc.body, 0, foreignTextNode, 36]
PASS Range 44: [xmlDoc, 0, xmlDoc, 0]
PASS Range 45: [xmlDoc, 1, xmlComment, 0]
PASS Range 46: [detachedTextNode, 0, detachedTextNode, 8]
PASS Range 47: [detachedForeignTextNode, 7, detachedForeignTextNode, 7]
PASS Range 48: [detachedForeignTextNode, 0, detachedForeignTextNode, 8]
PASS Range 49: [detachedXmlTextNode, 7, detachedXmlTextNode, 7]
PASS Range 50: [detachedXmlTextNode, 0, detachedXmlTextNode, 8]
PASS Range 51: [detachedComment, 3, detachedComment, 4]
PASS Range 52: [detachedComment, 5, detachedComment, 5]
PASS Range 53: [detachedForeignComment, 0, detachedForeignComment, 1]
PASS Range 54: [detachedForeignComment, 4, detachedForeignComment, 4]
PASS Range 55: [detachedXmlComment, 2, detachedXmlComment, 6]
PASS Range 56: [docfrag, 0, docfrag, 0]
PASS Range 57: [foreignDocfrag, 0, foreignDocfrag, 0]
PASS Range 58: [xmlDocfrag, 0, xmlDocfrag, 0]
